backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
python-dotenv>=1.0
# Uncomment for production:
# psycopg2-binary>=2.9    # PostgreSQL
# razorpay>=1.4           # Payment gateway
# boto3>=1.34             # AWS S3
# celery>=5.3             # Async tasks
# redis>=5.0              # Cache/broker
